+++ agape/src/error.rs
@@ -4,11 +4,14 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Debug, Error)]
 pub enum Error {
+    #[error("Unsupported image format, only png, jpeg and webp are supported")]
+    UnsupportedImageFormat,
+
     // Third party errors
     #[error(transparent)]
-    ReqwestError(#[from] reqwest::Error),
-    #[error(transparent)]
     ImageError(#[from] image::ImageError),
+    #[error("Io error: {0}")]
+    IoError(#[from] std::io::Error),
     #[error(transparent)]
     EventLoopError(#[from] winit::error::EventLoopError),
 }