+++ examples/vstack.rs
@@ -1,13 +1,18 @@
-use agape::{App, Color, vstack, widgets::Rect};
+use agape::widgets::{Image, Text};
+use agape::{App, vstack};
 
 fn main() -> Result<(), agape::Error> {
+    let image = Image::open("./examples/assets/PARTYNEXTDOOR Album Cover.jpg")?.fixed(250.0, 250.0);
+
     let vstack = vstack! {
-        Rect::new().fill().background_color(Color::BLACK),
-        Rect::new().fill().background_color(Color::BLACK),
+        image,
+        Text::new("PARTYNEXTDOOR").font_size(24),
+        Text::new("2013").font_size(12),
     }
     .spacing(12)
     .padding(12)
-    .background_color(Color::RED);
+    .fill()
+    .align_center();
 
     let app = App::new(vstack);
     app.run()